% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-dbplyr.R
\name{as_polars_lf.tbl_polarssql_connection}
\alias{as_polars_lf.tbl_polarssql_connection}
\alias{as_polars_df.tbl_polarssql_connection}
\alias{compute.tbl_polarssql_connection}
\alias{as.data.frame.tbl_polarssql_connection}
\alias{as_tibble.tbl_polarssql_connection}
\alias{collect.tbl_polarssql_connection}
\title{Compute results of a dbplyr query}
\usage{
\method{as_polars_lf}{tbl_polarssql_connection}(x, ..., cte = TRUE)

\method{as_polars_df}{tbl_polarssql_connection}(x, ..., n = Inf, cte = TRUE)

\method{compute}{tbl_polarssql_connection}(x, ..., n = Inf, cte = TRUE)

\method{as.data.frame}{tbl_polarssql_connection}(x, ..., n = Inf, cte = TRUE)

\method{as_tibble}{tbl_polarssql_connection}(x, ..., n = Inf, cte = TRUE)

\method{collect}{tbl_polarssql_connection}(x, ..., n = Inf, cte = TRUE)
}
\arguments{
\item{x}{A \link[=tbl_polarssql]{tbl_polarssql_connection} object.}

\item{...}{\itemize{
\item For \code{\link[=as_polars_lf]{as_polars_lf(<tbl_polarssql_connection>)}}: Ignored.
\item For \verb{as.data.frame(<tbl_polarssql_connection>)}: Other arguments passed to
\code{\link[polars:s3-as.data.frame]{as.data.frame(<polars_lazy_frame>)}}.
\item For \code{\link[tibble:as_tibble]{as_tibble(<tbl_polarssql_connection>)}} or
\code{\link[dplyr:compute]{collect(<tbl_polarssql_connection>)}}: Other arguments passed to
\code{\link[polars:s3-as_tibble]{as_tibble(<polars_lazy_frame>)}}.
\item For Other functions: Other arguments passed to \code{\link[=as_polars_df]{as_polars_df(<polars_lazy_frame>)}}.
}}

\item{cte}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Use common table expressions in the generated SQL?}

\item{n}{Number of rows to fetch. Defaults to \code{Inf}, meaning all rows.}
}
\description{
Compute results of a dbplyr query
}
\examples{
\dontshow{if (rlang::is_installed("dbplyr")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(dplyr, warn.conflicts = FALSE)
library(polars)

t <- tbl_polarssql(mtcars) |>
  filter(cyl == 4) |>
  mutate(cyl = sql("CAST (cyl AS Int64)"))

as_polars_lf(t)

as_polars_df(t, n = 1)

compute(t, n = Inf, engine = "streaming")

collect(t, n = 5, int64 = "character")

# Clean up
DBI::dbDisconnect(polarssql_default_connection())
\dontshow{\}) # examplesIf}
}
