% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarssql-package.R
\docType{package}
\name{polarssql-package}
\alias{polarssql-package}
\alias{_PACKAGE}
\title{polarssql: A 'polars' backend for 'DBI'.}
\description{
DBI-compliant interface to 'polars'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rpolars.github.io/r-polarssql/}
  \item \url{https://github.com/rpolars/r-polarssql}
  \item Report bugs at \url{https://github.com/rpolars/r-polarssql/issues}
}

}
\author{
\strong{Maintainer}: Tatsuya Shima \email{ts1s1andn@gmail.com}

}
\keyword{internal}
