% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\name{polarssql}
\alias{polarssql}
\title{polarssql driver}
\usage{
polarssql()
}
\description{
\code{\link[=polarssql]{polarssql()}} creates a DBI driver instance.
}
\examples{
\dontshow{if (polars::polars_info()$features$sql) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
polarssql()
\dontshow{\}) # examplesIf}
}
