% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_connection.R
\name{polarssql_default_connection}
\alias{polarssql_default_connection}
\title{Get the default connection}
\usage{
polarssql_default_connection()
}
\value{
A polarssql connection object
}
\description{
Get the default built-in connection.
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
polars::pl$LazyFrame(mtcars)$sink_parquet("mtcars.parquet")

query <- "SELECT * FROM read_parquet('mtcars.parquet') LIMIT 5"
con <- polarssql_default_connection()

polarssql_query(query, conn = con, result_type = "polars_df")
\dontshow{
setwd(.old_wd)
}
}
