% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{polarssql_register}
\alias{polarssql_register}
\alias{polarssql_unregister}
\title{Register data frames as tables}
\usage{
polarssql_register(conn, ..., overwrite = FALSE)

polarssql_unregister(conn, names)
}
\arguments{
\item{conn}{A polars sql connection, created by \code{\link[=polarssql]{polarssql()}}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Name-value pairs of \link{data.frame} like objects to register.}

\item{overwrite}{Should an existing registration be overwritten?}

\item{names}{Names of the tables to unregister.}
}
\value{
These functions are called for their side effect.
}
\description{
Register data frames as tables
}
\examples{
con <- dbConnect(polarssql())

polarssql_register(con, df1 = mtcars, df2 = mtcars)
con

polarssql_unregister(con, c("df1", "df2"))
con
}
