% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R, R/Connection.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dbCanConnect}
\alias{Id}
\alias{dbIsReadOnly}
\alias{dbQuoteLiteral}
\alias{dbUnquoteIdentifier}
\alias{dbGetQuery}
\alias{dbExecute}
\alias{dbReadTable}
\alias{dbCreateTable}
\alias{dbAppendTable}
\alias{dbListObjects}
\alias{dbWithTransaction}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DBI}{\code{\link[DBI]{dbAppendTable}}, \code{\link[DBI]{dbCanConnect}}, \code{\link[DBI]{dbCreateTable}}, \code{\link[DBI]{dbExecute}}, \code{\link[DBI]{dbGetQuery}}, \code{\link[DBI]{dbIsReadOnly}}, \code{\link[DBI]{dbListObjects}}, \code{\link[DBI]{dbQuoteLiteral}}, \code{\link[DBI]{dbReadTable}}, \code{\link[DBI]{dbUnquoteIdentifier}}, \code{\link[DBI]{dbWithTransaction}}, \code{\link[DBI]{Id}}}
}}

