% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-dbplyr.R
\name{as_polars_lf.tbl_polarssql_connection}
\alias{as_polars_lf.tbl_polarssql_connection}
\alias{as_polars_df.tbl_polarssql_connection}
\alias{compute.tbl_polarssql_connection}
\title{Compute results of a query}
\usage{
\method{as_polars_lf}{tbl_polarssql_connection}(x, ..., cte = TRUE)

\method{as_polars_df}{tbl_polarssql_connection}(x, ..., cte = TRUE)

\method{compute}{tbl_polarssql_connection}(x, ..., n = Inf, cte = TRUE)
}
\arguments{
\item{x}{A \link[=dbplyr-backend-polarssql]{tbl_polarssql_connection} object.}

\item{...}{Other arguments passed to \code{\link[=as_polars_df]{as_polars_df(<LazyFrame>)}}.}

\item{cte}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Use common table expressions in the generated SQL?}

\item{n}{Number of rows to fetch. Defaults to \code{Inf}, meaning all rows.}
}
\description{
Compute results of a query
}
\examples{
\dontshow{if (polars::pl$polars_info()$features$sql && rlang::is_installed("dbplyr")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(dplyr, warn.conflicts = FALSE)
library(polars)

t <- tbl_polarssql(mtcars) |>
  filter(cyl == 4)

as_polars_lf(t)

as_polars_df(t,  n_rows = 1)

compute(t, n = 1) # Equivalent to `as_polars_df(t, n_rows = 1)`
\dontshow{\}) # examplesIf}
}
\seealso{
\link{dbplyr-backend-polarssql}
}
