% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R, R/Connection.R,
%   R/Result.R, R/dbClearResult_Result.R, R/dbConnect_Driver.R,
%   R/dbDisconnect_Connectsion.R, R/dbFetch_Result.R, R/dbHasCompleted_Result.R,
%   R/dbIsValid_Connection.R, R/dbIsValid_Driver.R, R/dbIsValid_Result.R,
%   R/dbSendQuery_Connection_character.R, R/dbWriteTable_Connection_character.R,
%   R/show_Connection.R, R/show_Driver.R, R/show_Result.R
\docType{class}
\name{DBI}
\alias{DBI}
\alias{polars_sql_driver-class}
\alias{polars_sql_connection-class}
\alias{polars_sql_result-class}
\alias{dbClearResult_polars_sql_result}
\alias{dbClearResult,polars_sql_result-method}
\alias{dbConnect_polars_sql_driver}
\alias{dbConnect,polars_sql_driver-method}
\alias{dbDisconnect_polars_sql_connection}
\alias{dbDisconnect,polars_sql_connection-method}
\alias{dbFetch_polars_sql_result}
\alias{dbFetch,polars_sql_result-method}
\alias{dbHasCompleted_polars_sql_result}
\alias{dbHasCompleted,polars_sql_result-method}
\alias{dbIsValid_polars_sql_connection}
\alias{dbIsValid,polars_sql_connection-method}
\alias{dbIsValid_polars_sql_driver}
\alias{dbIsValid,polars_sql_driver-method}
\alias{dbIsValid_polars_sql_result}
\alias{dbIsValid,polars_sql_result-method}
\alias{dbSendQuery_polars_sql_connection_character}
\alias{dbSendQuery,polars_sql_connection,character-method}
\alias{dbWriteTable_polars_sql_connection_character_data.frame}
\alias{dbWriteTable,polars_sql_connection,character,data.frame-method}
\alias{show_polars_sql_connection}
\alias{show,polars_sql_connection-method}
\alias{show_polars_sql_driver}
\alias{show,polars_sql_driver-method}
\alias{show_polars_sql_result}
\alias{show,polars_sql_result-method}
\title{DBI methods}
\usage{
\S4method{dbClearResult}{polars_sql_result}(res, ...)

\S4method{dbConnect}{polars_sql_driver}(drv, ...)

\S4method{dbDisconnect}{polars_sql_connection}(conn, ...)

\S4method{dbFetch}{polars_sql_result}(res, n = -1, ...)

\S4method{dbHasCompleted}{polars_sql_result}(res, ...)

\S4method{dbIsValid}{polars_sql_connection}(dbObj, ...)

\S4method{dbIsValid}{polars_sql_driver}(dbObj, ...)

\S4method{dbIsValid}{polars_sql_result}(dbObj, ...)

\S4method{dbSendQuery}{polars_sql_connection,character}(conn, statement, ...)

\S4method{dbWriteTable}{polars_sql_connection,character,data.frame}(conn, name, value, overwrite = FALSE, ...)

\S4method{show}{polars_sql_connection}(object)

\S4method{show}{polars_sql_driver}(object)

\S4method{show}{polars_sql_result}(object)
}
\arguments{
\item{res}{An object inheriting from \linkS4class{DBIResult}.}

\item{...}{Other arguments passed on to methods.}

\item{drv}{an object that inherits from \linkS4class{DBIDriver},
or an existing \linkS4class{DBIConnection}
object (in order to clone an existing connection).}

\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}

\item{statement}{a character string containing SQL.}

\item{name}{The table name, passed on to \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}. Options are:
\itemize{
\item a character string with the unquoted DBMS table name,
e.g. \code{"table_name"},
\item a call to \code{\link[DBI:Id]{Id()}} with components to the fully qualified table name,
e.g. \code{Id(schema = "my_schema", table = "table_name")}
\item a call to \code{\link[DBI:SQL]{SQL()}} with the quoted and fully qualified table name
given verbatim, e.g. \code{SQL('"my_schema"."table_name"')}
}}

\item{value}{a \link{data.frame} (or coercible to data.frame).}

\item{overwrite}{Allow overwriting the destination table. Cannot be
\code{TRUE} if \code{append} is also \code{TRUE}.}

\item{object}{Any R object}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package.
}
