% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarssql-package.R
\docType{package}
\name{polarssql-package}
\alias{polarssql-package}
\alias{_PACKAGE}
\title{polarssql: A 'polars' backend for 'DBI'.}
\description{
DBI-compliant interface to 'polars'.
}
\author{
\strong{Maintainer}: Tatsuya Shima \email{ts1s1andn@gmail.com}

}
\keyword{internal}
