% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarssql-package.R
\docType{package}
\name{polarssql-package}
\alias{polarssql}
\alias{polarssql-package}
\title{polarssql: A 'polars' backend for 'DBI'.}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Tatsuya Shima \email{ts1s1andn@gmail.com}

}
\keyword{internal}
