% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{polarssql_query}
\alias{polarssql_query}
\title{Execute SQL query}
\usage{
polarssql_query(sql, conn = polarssql_default_connection())
}
\arguments{
\item{sql}{A SQL string.}

\item{conn}{A polarssql connection, created by \code{\link[=polarssql]{polarssql()}}.
Use the default built-in connection by default.}
}
\value{
\link[polars:LazyFrame_class]{polars LazyFrame}
}
\description{
Execute SQL query
}
\examples{
\dontshow{if (polars::pl$polars_info()$features$sql) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
polarssql_register(mtcars = mtcars)

query <- "SELECT * FROM mtcars LIMIT 5"

# Returns a polars LazyFrame
polarssql_query(query)

# Clean up
polarssql_unregister("mtcars")
\dontshow{\}) # examplesIf}
}
