% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{polarssql_register}
\alias{polarssql_register}
\alias{polarssql_unregister}
\title{Register data frames as tables}
\usage{
polarssql_register(
  ...,
  .conn = polarssql_default_connection(),
  .overwrite = FALSE
)

polarssql_unregister(names, conn = polarssql_default_connection())
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Name-value pairs of \link{data.frame} like objects to register.}

\item{.conn, conn}{A polarssql connection, created by \code{\link[=polarssql]{polarssql()}}.
Use the default built-in connection by default.}

\item{.overwrite}{Should an existing registration be overwritten?}

\item{names}{Names of the tables to unregister.}
}
\value{
These functions are called for their side effect.
}
\description{
Register data frames as tables
}
\examples{
\dontshow{if (polars::pl$polars_info()$features$sql) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
con <- dbConnect(polarssql())

polarssql_register(df1 = mtcars, df2 = mtcars, .conn = con)
con

polarssql_unregister(c("df1", "df2"), conn = con)
con
\dontshow{\}) # examplesIf}
}
